// src/api/requests/models/NotificationEntity.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type NotificationEntity = {
  id: string;
  userId: string;
  businessId: string | null;
  title: string;
  message: string;
  isSeen: boolean;
  notificationType: NotificationEntity.notificationType;
  metaData: Record<string, any>;
  createdAt: string;
};

export namespace NotificationEntity {
  export enum notificationType {
    NEWS = 'NEWS',
    BOOKING_CONFIRMED = 'BOOKING_CONFIRMED',
    BOOKING_DRIVER_ON_THE_WAY = 'BOOKING_DRIVER_ON_THE_WAY',
    BOOKING_DRIVER_ARRIVED = 'BOOKING_DRIVER_ARRIVED',
    BOOKING_CANCELLED = 'BOOKING_CANCELLED',
    BOOKING_AUTO_CANCELLED = 'BOOKING_AUTO_CANCELLED',
    BOOKING_COMPLETED = 'BOOKING_COMPLETED',
    BOOKING_SCHEDULE = 'BOOKING_SCHEDULE',
    BOOKING_PAYMENT_SUCCESS = 'BOOKING_PAYMENT_SUCCESS',
    BOOKING_PAYMENT_FAILURE = 'BOOKING_PAYMENT_FAILURE',
    BOOKING_RIDE_TOMORROW = 'BOOKING_RIDE_TOMORROW',
    BOOKING_RIDE_IN_HOUR = 'BOOKING_RIDE_IN_HOUR',
    BOOKING_WARNING_CANCEL = 'BOOKING_WARNING_CANCEL',
    USER_COUPON_REMINDER = 'USER_COUPON_REMINDER',
    FIRST_RIDE = 'FIRST_RIDE',
    COMPLETED_RIDE = 'COMPLETED_RIDE',
    COUPON_REMINDER = 'COUPON_REMINDER',
    BOOKING_REMINDER = 'BOOKING_REMINDER',
    SUCCESSFUL_REFERRAL = 'SUCCESSFUL_REFERRAL',
    REFERRAL_BONUS_UNLOCKED = 'REFERRAL_BONUS_UNLOCKED',
    SUBSCRIPTION_RENEWAL_FAILED = 'SUBSCRIPTION_RENEWAL_FAILED',
    SUBSCRIPTION_RENEWAL_SUCCESSFULLY = 'SUBSCRIPTION_RENEWAL_SUCCESSFULLY',
    SUBSCRIPTION_CANCELLED = 'SUBSCRIPTION_CANCELLED',
    SUBSCRIPTION_RESUBSCRIBE = 'SUBSCRIPTION_RESUBSCRIBE',
    SUBSCRIPTION_EXPIRED = 'SUBSCRIPTION_EXPIRED',
    SUBSCRIPTION_REMINDER = 'SUBSCRIPTION_REMINDER',
    SUBMITTED_ESSAY_TEST = 'SUBMITTED_ESSAY_TEST',
    GRADED_ESSAY_TEST = 'GRADED_ESSAY_TEST',
    NEW_COMMENT_POST = 'NEW_COMMENT_POST',
    LIKE_COMMENT_POST = 'LIKE_COMMENT_POST',
    LIKE_POST = 'LIKE_POST',
  }
}